Full Update
Profile
Users
Manuals
ManualItems
ManualItemContent
Update on all of them

+++ app/Http/Controllers/ManualItemContentController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Models\ManualItemContent;
 use App\Models\ManualsItem;
+use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Spatie\Permission\Models\Permission;
 
 class ManualItemContentController extends Controller
 {
@@ -25,28 +29,15 @@ class ManualItemContentController extends Controller
         $validate = $request->validate([
             'manual_name' => 'string',
             'type' => 'string',
-            'files' => 'array|max:5', // Limit to 5 uploads
-            'files.*' => 'file|mimes:jpg,jpeg,png,pdf,doc,docx,webp,xlsx|max:40960', // Validation rules for each file
+            'files' => 'array|max:10', // Limit to 5 uploads
+            'files.*' => 'file|mimes:pdf|max:40960', // Validation rules for each file
         ]);
-        if (!empty($validate['type']) && $validate['type'] == 'Folder') {
-            $manualItemContent::create([
-                'miid' => uuid_create(UUID_TYPE_DEFAULT),
-                'manual_uid' => $request->manual_uid,
-                'manual_iid' => $request->id,
-                'name' => $request->manual_name,
-                'file_size' => '0MB', // Size in byte
-                'file_type' => 'Folder',
-            ]);
-            return redirect(route('manual.items.content.index', $request->id))->with('success', 'Folder Created');
-        } else {
+
             foreach ($request->file('file') as $file) {
-//                $fileName = uniqid() . '.' . $file->getClientOriginalExtension();
                 $fileName = $file->getClientOriginalName();
-                $path = $file->storeAs('public/uploads/items/content', $fileName);
-
+                $fileNameUnique = Str::random(4) . '_' . $fileName;
+                $path = Storage::disk('privateSubManualContent')->putFileAs('', $file, $fileNameUnique);
                 $fileSize = $file->getSize(); // Get file size in bytes
-//                $fileSizeInMB = $fileSize / 1024 / 1024; // Convert bytes to MB
-
                 $mimeType = $file->getClientMimeType(); // Get MIME type
 
                 try {
@@ -55,11 +46,10 @@ class ManualItemContentController extends Controller
                         'micd' => uuid_create(UUID_TYPE_DEFAULT),
                         'manual_uid' => $request->id,
                         'manual_iid' => $request->manual_uid,
-                        'name' => $fileName,
+                        'name' => Str::beforeLast($fileName, '.pdf'),
                         'link' => $path,
                         'file_size' => $fileSize, // Size in byte
-                        'file_type' => $mimeType,
-                        // Add additional fields (e.g., path)
+                        'file_type' => $mimeType, //Application type
                     ];
                     // Use your model to save data to the database (replace with your model logic)
                     $manualItemContent::insert($fileData);
@@ -70,10 +60,19 @@ class ManualItemContentController extends Controller
                     ]);
                 }
             }
+
+    }
+
+    private function giveAllPermissions($permission)
+    {
+        $roles = Role::whereIn('name', ['Admin', 'Librarian'])->get();
+
+        foreach ($roles as $role) {
+            // Append the permission to all users with the role
+            foreach ($role->users as $user) {
+                $user->givePermissionTo($permission);
+            }
         }
-        return response()->json([
-            'message' => 'File uploaded successfully!',
-        ]);
     }
 
     /**
@@ -113,17 +112,24 @@ class ManualItemContentController extends Controller
      */
     public function destroy($id, $ids, ManualItemContent $manualItemContent)
     {
-        $path = $manualItemContent::where('micd', $ids)->first();
-        if (!empty($path['file_type']) && $path['file_type'] != 'Folder') {
-            if (!empty($path['link'])) {
-                Storage::delete($path['link']);
-                $manualItemContent::where('micd', $ids)->delete();
-                return redirect(route('manual.items.content.index', $id, $ids))->with('success', $path['name'].' File Deleted');
-            }else{
-                $manualItemContent::where('micd', $ids)->delete();
-                return redirect(route('manual.items.content.index', $id, $ids))->with('success', $path['name'].' File Deleted');
+        if (Auth::user()->can('can destroy')) {
+            $path = $manualItemContent::where('micd', $ids)->first();
+            if (!empty($path['file_type']) && $path['file_type'] != 'Folder') {
+                if (!empty($path['link'])) {
+                    if (Storage::disk('privateSubManualContent')->exists($path['link'])) {
+                        Storage::disk('privateSubManualContent')->delete($path['link']);
+                        $manualItemContent::where('micd', $ids)->delete();
+                        return redirect(route('manual.items.content.index', $id, $ids))->with('success', $path['name'] . ' File Deleted');
+                    } else {
+                        return response()->json(['error' => 'File not found'], 404);
+                    }
+                } else {
+                    $manualItemContent::where('micd', $ids)->delete();
+                    return redirect(route('manual.items.content.index', $id, $ids))->with('success', $path['name'] . ' File Deleted');
+                }
             }
+        } else {
+            return response()->json(['error' => 'You do not have permission to delete files'], 404);
         }
-//        return redirect(route('manual.items.content.index', $id, $ids))->with('success', 'Folder or File Deleted');
     }
 }
